sound database 추가 / 기본 crud 코드 작성 / face, record,video datatype 수정 /

+++ models/record.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, ForeignKey, Integer, String, TIMESTAMP
+from sqlalchemy import Column, ForeignKey, Integer, Float, String, TIMESTAMP
 from sqlalchemy.orm import relationship
 
 from config.database import Base
@@ -12,14 +12,14 @@ class Record(Base):
     created_at = Column(TIMESTAMP)
 
     label = Column(String(400))
-    anger_score = Column(Integer)
-    scorn_score = Column(Integer)
-    disgust_score = Column(Integer)
-    happy_score = Column(Integer)
-    neutral_score = Column(Integer)
-    sad_score = Column(Integer)
-    surprised_score = Column(Integer)
-    voice_score = Column(Integer)
+    anger_score = Column(Float)
+    scorn_score = Column(Float)
+    disgust_score = Column(Float)
+    happy_score = Column(Float)
+    neutral_score = Column(Float)
+    sad_score = Column(Float)
+    surprised_score = Column(Float)
+    voice_score = Column(Float)
 
     user = relationship("User", back_populates="records")
     video = relationship("Video", back_populates="record")
