+++ schemas/face_schema.py
@@ -7,14 +7,14 @@ class FaceBase(BaseModel):
 
 
 class FaceCreate(FaceBase):
-    anger_score: int
-    scorn_score: int
-    disgust_score: int
-    happy_score: int
-    neutral_score: int
-    sad_score: int
-    surprised_score: int
-    voice_score: int
+    anger_score: float
+    scorn_score: float
+    disgust_score: float
+    happy_score: float
+    neutral_score: float
+    sad_score: float
+    surprised_score: float
+    voice_score: float
     started_at: datetime
     ended_at: datetime
